refactor: extract SupportDataActionPresenter.swift from SupportDataTypes.swift
Splits the `SupportDataActionPresenter` @MainActor class (~87 lines,
UI-adjacent orchestration) out of SupportDataTypes.swift so that file
holds only the completion/failure data types.

Byte-identical move of lines 46-132. Only imports Foundation.

SupportDataTypes.swift: 133 → 45 lines. Tests: 667.
